backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
google-generativeai==0.7.2
python-multipart==0.0.9
numpy==1.26.4
huggingface-hub==0.25.2

